requirements.txt
torch==1.13.1
torchvision==0.14.1
networkx==3.0
numpy==1.22.4
PyYAML==6.0
scikit-image==0.19.1
scipy==1.7.3
Shapely==1.8.0